+++ pkg/webui/console/views/device-general-settings/network-server-form/index.js
@@ -22,6 +22,7 @@ import Select from '@ttn-lw/components/select'
 import Form from '@ttn-lw/components/form'
 import Notification from '@ttn-lw/components/notification'
 import Checkbox from '@ttn-lw/components/checkbox'
+import toast from '@ttn-lw/components/toast'
 
 import PhyVersionInput from '@console/components/phy-version-input'
 import MacSettingsSection from '@console/components/mac-settings-section'
@@ -29,6 +30,7 @@ import MacSettingsSection from '@console/components/mac-settings-section'
 import { NsFrequencyPlansSelect } from '@console/containers/freq-plans-select'
 import DevAddrInput from '@console/containers/dev-addr-input'
 
+import { isBackend, getBackendErrorName } from '@ttn-lw/lib/errors/utils'
 import tooltipIds from '@ttn-lw/lib/constants/glossary-ids'
 import diff from '@ttn-lw/lib/diff'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
@@ -59,14 +61,38 @@ const defaultValues = {
 }
 
 const NetworkServerForm = React.memo(props => {
-  const { device, onSubmit, onSubmitSuccess, mayEditKeys, mayReadKeys } = props
-  const { multicast = false, supports_join = false, supports_class_b = false } = device
+  const {
+    device,
+    onSubmit,
+    onSubmitSuccess,
+    mayEditKeys,
+    mayReadKeys,
+    getDefaultMacSettings,
+  } = props
+  const {
+    multicast = false,
+    supports_join = false,
+    supports_class_b = false,
+    supports_class_c = false,
+  } = device
 
   const isABP = isDeviceABP(device)
   const isMulticast = isDeviceMulticast(device)
   const isJoinedOTAA = isDeviceOTAA(device) && isDeviceJoined(device)
 
   const formRef = React.useRef(null)
+
+  const [macSettings, setMacSettings] = React.useState({})
+  const [freqPlan, setFreqPlan] = React.useState(device.frequency_plan_id)
+  const freqPlanRef = React.useRef()
+  const handleFreqPlanChange = React.useCallback(band => {
+    setFreqPlan(band.value)
+  }, [])
+  const [phyVersion, setPhyVersion] = React.useState(device.lorawan_phy_version)
+  const phyVersionRef = React.useRef()
+  const handlePhyVersionChange = React.useCallback(version => {
+    setPhyVersion(version)
+  }, [])
 
   const [error, setError] = React.useState('')
 
@@ -79,6 +105,51 @@ const NetworkServerForm = React.memo(props => {
 
     setClassB(checked)
   }, [])
+  const [isClassC, setClassC] = React.useState(supports_class_c)
+  const handleClassCChange = React.useCallback(evt => {
+    const { checked } = evt.target
+
+    setClassC(checked)
+  }, [])
+
+  React.useEffect(() => {
+    const getMacSettings = async (freqPlan, phyVersion) => {
+      try {
+        const settings = await getDefaultMacSettings(freqPlan, phyVersion)
+        setMacSettings(settings)
+      } catch (err) {
+        if (isBackend(err) && getBackendErrorName(err) === 'no_band_version') {
+          toast({
+            type: toast.types.ERROR,
+            message: messages.fpNotFoundError,
+            messageValues: {
+              lorawanVersion,
+              freqPlan,
+              code: msg => <code>{msg}</code>,
+            },
+          })
+        } else {
+          toast({
+            type: toast.types.ERROR,
+            message: messages.macSettingsError,
+            messageValues: {
+              freqPlan,
+              code: msg => <code>{msg}</code>,
+            },
+          })
+        }
+      }
+    }
+
+    if (freqPlan && phyVersion) {
+      if (freqPlanRef.current !== freqPlan || phyVersionRef.current !== phyVersion) {
+        freqPlanRef.current = freqPlan
+        phyVersionRef.current = phyVersion
+
+        getMacSettings(freqPlan, phyVersion)
+      }
+    }
+  }, [freqPlan, getDefaultMacSettings, lorawanVersion, phyVersion])
 
   const initialActivationMode = supports_join
     ? ACTIVATION_MODES.OTAA
@@ -106,12 +177,13 @@ const NetworkServerForm = React.memo(props => {
           _device_classes: { class_b: device.supports_class_b, class_c: device.supports_class_c },
           mac_settings: {
             ...defaultValues.mac_settings,
+            ...macSettings,
             ...device.mac_settings,
           },
         },
         { context: validationContext },
       ),
-    [device, initialActivationMode, validationContext],
+    [device, initialActivationMode, macSettings, validationContext],
   )
 
   const onFormSubmit = React.useCallback(
@@ -252,11 +324,13 @@ const NetworkServerForm = React.memo(props => {
         lorawanVersion={lorawanVersion}
         required
         tooltipId={tooltipIds.REGIONAL_PARAMETERS}
+        onChange={handlePhyVersionChange}
       />
       <NsFrequencyPlansSelect
         name="frequency_plan_id"
         required
         tooltipId={tooltipIds.FREQUENCY_PLAN}
+        onChange={handleFreqPlanChange}
       />
       <Form.Field
         title={sharedMessages.lorawanClassCapabilities}
@@ -271,7 +345,11 @@ const NetworkServerForm = React.memo(props => {
           label={sharedMessages.supportsClassB}
           onChange={handleClassBChange}
         />
-        <Checkbox name="class_c" label={sharedMessages.supportsClassC} />
+        <Checkbox
+          name="class_c"
+          label={sharedMessages.supportsClassC}
+          onChange={handleClassCChange}
+        />
       </Form.Field>
       <Form.Field
         title={sharedMessages.activationMode}
@@ -339,7 +417,11 @@ const NetworkServerForm = React.memo(props => {
           )}
         </>
       )}
-      <MacSettingsSection activationMode={initialActivationMode} isClassB={isClassB} />
+      <MacSettingsSection
+        activationMode={initialActivationMode}
+        isClassB={isClassB}
+        isClassC={isClassC}
+      />
       <SubmitBar>
         <Form.Submit component={SubmitButton} message={sharedMessages.saveChanges} />
       </SubmitBar>
@@ -349,6 +431,7 @@ const NetworkServerForm = React.memo(props => {
 
 NetworkServerForm.propTypes = {
   device: PropTypes.device.isRequired,
+  getDefaultMacSettings: PropTypes.func.isRequired,
   mayEditKeys: PropTypes.bool.isRequired,
   mayReadKeys: PropTypes.bool.isRequired,
   onSubmit: PropTypes.func.isRequired,